Fix apostrophes in text
