+++ crates/html-macro-test/src/tests.rs
@@ -1,0 +1,3 @@
+mod all_tests;
+mod text;
+mod ui;
