+++ crates/html-macro-test/src/tests/ui/should_be_self_closing_tag.rs
