+++ crates/html-macro-test/src/tests/ui/should_be_self_closing_tag.stderr
@@ -3,6 +3,3 @@ error: br is a self closing tag. Try "<br>" or "<br />"
    |
 10 |         <br></br>
    |               ^^
-
-error: aborting due to previous error
-
