+++ crates/html-macro-test/src/tests/ui/wrong_closing_tag.rs
