--- crates/html-macro-ui/README.md
@@ -1,3 +1,0 @@
-# html-macro-ui
-
-Tests that verify the compile time errors ofour html! macro.
